+++ crates/ironclaw_operator/src/llm_admin/active_model.rs
@@ -12,7 +12,7 @@
 use std::sync::Arc;
 
 use ironclaw_llm::LlmProvider;
-use ironclaw_product::ActiveModelReader;
+use ironclaw_product_contracts::llm_config::ActiveModelReader;
 
 /// [`ActiveModelReader`] over the runtime's live primary provider handle.
 pub struct ProviderActiveModelReader {
