+++ crates/ironclaw_operator/src/llm_admin/provider_admin.rs
@@ -7,6 +7,7 @@
 
 use std::path::PathBuf;
 
+use ironclaw_product_contracts::llm_config::LlmProbeRequest;
 use ironclaw_product_contracts::operator_llm::{
     DetectedEnvLlm, EXAMPLE_OVERLAY_PROVIDER_ID, ProviderMenuEntry, ProviderProbeOutcome,
     RebornModelRoutesState, RebornProviderInfo, RebornProviderList, RebornProviderMetadata,
@@ -425,7 +426,7 @@ impl RebornProviderAdmin {
             }
         })?;
         let base_url = candidate_probe_base_url(definition);
-        let request = ironclaw_product::LlmProbeRequest {
+        let request = LlmProbeRequest {
             adapter: provider_protocol_wire_name(definition.protocol),
             base_url,
             provider_id: provider_id.to_string(),
