+++ crates/ironclaw_product/src/reborn_services.rs
@@ -22,6 +22,18 @@ use ironclaw_product_contracts::channel_config::ChannelConfigProductService;
 use ironclaw_product_contracts::lifecycle_service::{
     LifecycleProductContext, LifecycleProductService, LifecycleProductSurfaceContext,
 };
+use ironclaw_product_contracts::llm_config::{
+    ActiveModelReader, CodexLoginStart, LlmConfigService, LlmConfigSnapshot, LlmModelsResult,
+    LlmProbeRequest, LlmProbeResult, NearAiLoginRequest, NearAiLoginStart,
+    NearAiWalletLoginRequest, NearAiWalletLoginResult, UpsertLlmProviderRequest,
+};
+use ironclaw_product_contracts::operator_service::{
+    OperatorLogsService, OperatorServiceLifecycleService, OperatorStatusService, RebornLogEntry,
+    RebornLogQueryRequest, RebornLogQueryResponse, RebornOperatorStatusCheck,
+    RebornOperatorStatusResponse, RebornOperatorStatusSeverity, RebornOperatorStatusState,
+    RebornServiceLifecycleAction, RebornServiceLifecycleRequest, RebornServiceLifecycleResponse,
+    RebornServiceLifecycleState, normalize_operator_log_context_value,
+};
 use ironclaw_product_contracts::operator_tools::{
     RebornOperatorToolCatalog, RebornOperatorToolInfo,
 };
@@ -170,13 +182,7 @@ pub use fs_browse::{
 };
 pub use ironclaw_product_contracts::package_lifecycle::ChannelConnectStrategy as RebornChannelConnectStrategy;
 pub use lifecycle_setup::EXTENSION_SETUP_VIEW;
-pub use llm_config::{
-    ActiveModelReader, CodexLoginStart, LLM_CONFIG_VIEW, LlmActiveSelection, LlmConfigService,
-    LlmConfigServiceError, LlmConfigSnapshot, LlmModelsResult, LlmProbeRequest, LlmProbeResult,
-    LlmProviderView, NearAiAuthProvider, NearAiLoginRequest, NearAiLoginStart,
-    NearAiWalletLoginRequest, NearAiWalletLoginResult, SetActiveLlmRequest,
-    UpsertLlmProviderRequest,
-};
+pub use llm_config::LLM_CONFIG_VIEW;
 pub use log_views::{LOGS_VIEW, OPERATOR_LOGS_VIEW};
 pub use operator_command_views::{
     OPERATOR_DIAGNOSTICS_VIEW, OPERATOR_SETUP_VIEW, OPERATOR_STATUS_VIEW,
@@ -232,8 +238,7 @@ pub use types::{
     RebornExtensionRegistryResponse, RebornExtensionSetupField, RebornExtensionSetupSecret,
     RebornExtensionSurface, RebornGetRunStateRequest, RebornGetRunStateResponse,
     RebornGlobalAutoApproveRequest, RebornGlobalAutoApproveResponse, RebornListAutomationsResponse,
-    RebornListThreadsResponse, RebornLogEntry, RebornLogLevel, RebornLogQueryRequest,
-    RebornLogQueryResponse, RebornOperatorArea, RebornOperatorCommandPlaneResponse,
+    RebornListThreadsResponse, RebornOperatorArea, RebornOperatorCommandPlaneResponse,
     RebornOperatorConfigDiagnostic, RebornOperatorConfigDiagnosticSeverity,
     RebornOperatorConfigEntry, RebornOperatorConfigGetResponse, RebornOperatorConfigListResponse,
     RebornOperatorConfigSetProductRequest, RebornOperatorConfigSetRequest,
@@ -241,8 +246,7 @@ pub use types::{
     RebornOperatorLogsQuery, RebornOperatorServiceLifecycleAction,
     RebornOperatorServiceLifecycleRequest, RebornOperatorSetupRequest, RebornOperatorSetupResponse,
     RebornOperatorSetupStatus, RebornOperatorSetupStep, RebornOperatorSetupStepStatus,
-    RebornOperatorStatusCheck, RebornOperatorStatusResponse, RebornOperatorStatusSeverity,
-    RebornOperatorStatusState, RebornOperatorSurfaceStatus, RebornOutboundDeliveryModality,
+    RebornOperatorSurfaceStatus, RebornOutboundDeliveryModality,
     RebornOutboundDeliveryTargetCapabilities, RebornOutboundDeliveryTargetChannel,
     RebornOutboundDeliveryTargetDescription, RebornOutboundDeliveryTargetDisplayName,
     RebornOutboundDeliveryTargetId, RebornOutboundDeliveryTargetListResponse,
@@ -250,14 +254,12 @@ pub use types::{
     RebornOutboundDeliveryTargetSummary, RebornOutboundPreferencesResponse,
     RebornProductCommandInfo, RebornProductCommandListResponse,
     RebornRenameAutomationProductRequest, RebornResolveGateResponse, RebornResumeGateResponse,
-    RebornRetryRunResponse, RebornServiceLifecycleAction, RebornServiceLifecycleRequest,
-    RebornServiceLifecycleResponse, RebornServiceLifecycleState,
-    RebornSetOutboundPreferencesRequest, RebornSetupExtensionResponse, RebornSkillActionResponse,
-    RebornSkillContentResponse, RebornSkillInfo, RebornSkillListResponse,
-    RebornSkillSearchResponse, RebornSkillSourceKind, RebornSkillTrustLevel,
-    RebornStreamEventsRequest, RebornStreamEventsResponse, RebornSubmitTurnResponse,
-    RebornTimelineRequest, RebornTimelineResponse, RebornTraceHoldAuthorizeProductRequest,
-    RebornVendorAuthAccounts,
+    RebornRetryRunResponse, RebornSetOutboundPreferencesRequest, RebornSetupExtensionResponse,
+    RebornSkillActionResponse, RebornSkillContentResponse, RebornSkillInfo,
+    RebornSkillListResponse, RebornSkillSearchResponse, RebornSkillSourceKind,
+    RebornSkillTrustLevel, RebornStreamEventsRequest, RebornStreamEventsResponse,
+    RebornSubmitTurnResponse, RebornTimelineRequest, RebornTimelineResponse,
+    RebornTraceHoldAuthorizeProductRequest, RebornVendorAuthAccounts,
 };
 pub use views::{ProductView, UnavailableRebornViewProvider};
 
@@ -563,9 +565,6 @@ const OPERATOR_LOGS_DEFAULT_LIMIT: u32 = 100;
 const OPERATOR_LOGS_MAX_LIMIT: u32 = 500;
 const OPERATOR_LOGS_CURSOR_MAX_BYTES: usize = 512;
 const OPERATOR_LOGS_TARGET_MAX_BYTES: usize = 256;
-const OPERATOR_LOGS_CONTEXT_MAX_BYTES: usize = 256;
-const OPERATOR_LOG_CONTEXT_TRUNCATED_SUFFIX: &str = " ... [truncated]";
-
 const NOTICE_BLOCKED_APPROVAL: &str = "An approval gate is open on this thread — resolve it (approve or deny) before continuing, then resend your message.";
 const NOTICE_BLOCKED_AUTH: &str = "An authentication gate is open on this thread — complete authentication before continuing, then resend your message.";
 const NOTICE_BUSY_GENERIC: &str = "Ironclaw is still working on a previous message — resend yours once the current task finishes.";
@@ -579,7 +578,7 @@ fn command_result_field(label: &str, value: impl Into<String>) -> CommandResultF
     }
 }
 
-fn model_command_view(title: &str, snapshot: &llm_config::LlmConfigSnapshot) -> CommandResultView {
+fn model_command_view(title: &str, snapshot: &LlmConfigSnapshot) -> CommandResultView {
     let mut fields = Vec::new();
     let mut lines = Vec::new();
     match &snapshot.active {
@@ -723,14 +722,6 @@ impl ChannelConnectionService for StaticChannelConnectionService {
     }
 }
 
-#[async_trait]
-pub trait OperatorStatusService: Send + Sync {
-    async fn status(
-        &self,
-        caller: ProductSurfaceCaller,
-    ) -> Result<RebornOperatorStatusResponse, ProductSurfaceError>;
-}
-
 #[derive(Debug, Clone)]
 pub struct StaticOperatorStatusService {
     response: RebornOperatorStatusResponse,
@@ -765,15 +756,6 @@ impl OperatorStatusService for UnsupportedOperatorStatusService {
     }
 }
 
-#[async_trait]
-pub trait OperatorLogsService: Send + Sync {
-    async fn query_logs(
-        &self,
-        caller: ProductSurfaceCaller,
-        request: RebornLogQueryRequest,
-    ) -> Result<RebornLogQueryResponse, ProductSurfaceError>;
-}
-
 #[derive(Debug, Default)]
 pub struct UnsupportedOperatorLogsService;
 
@@ -786,15 +768,6 @@ impl OperatorLogsService for UnsupportedOperatorLogsService {
     ) -> Result<RebornLogQueryResponse, ProductSurfaceError> {
         Err(operator_surface_unavailable())
     }
-}
-
-#[async_trait]
-pub trait OperatorServiceLifecycleService: Send + Sync {
-    async fn control_service(
-        &self,
-        caller: ProductSurfaceCaller,
-        request: RebornServiceLifecycleRequest,
-    ) -> Result<RebornServiceLifecycleResponse, ProductSurfaceError>;
 }
 
 #[derive(Debug, Default)]
@@ -7138,36 +7111,12 @@ fn bounded_operator_logs_context_string(value: Option<String>) -> Option<String>
     })
 }
 
-pub fn normalize_operator_log_context_value(value: &str) -> String {
-    truncate_utf8_with_suffix(value, OPERATOR_LOGS_CONTEXT_MAX_BYTES)
-}
-
 fn truncate_utf8_to_bytes(value: &str, max_bytes: usize) -> String {
     let mut end = max_bytes.min(value.len());
     while end > 0 && !value.is_char_boundary(end) {
         end -= 1;
     }
     value[..end].to_string()
-}
-
-fn truncate_utf8_with_suffix(value: &str, max_bytes: usize) -> String {
-    if value.len() <= max_bytes {
-        return value.to_string();
-    }
-
-    if max_bytes <= OPERATOR_LOG_CONTEXT_TRUNCATED_SUFFIX.len() {
-        return OPERATOR_LOG_CONTEXT_TRUNCATED_SUFFIX[..max_bytes].to_string();
-    }
-
-    let mut end = max_bytes - OPERATOR_LOG_CONTEXT_TRUNCATED_SUFFIX.len();
-    while end > 0 && !value.is_char_boundary(end) {
-        end -= 1;
-    }
-
-    let mut truncated = String::with_capacity(max_bytes);
-    truncated.push_str(&value[..end]);
-    truncated.push_str(OPERATOR_LOG_CONTEXT_TRUNCATED_SUFFIX);
-    truncated
 }
 
 fn product_agent_bound_caller_from_webui(
