+++ crates/ironclaw_product/src/reborn_services/types.rs
@@ -10,6 +10,13 @@ use ironclaw_host_api::turn::{
     TurnStatus,
 };
 use ironclaw_loop_contracts::LoopModelUsage;
+// The operator control plane's port vocabulary is declared at the product
+// boundary (`ironclaw_product_contracts::operator_service`); product's own
+// operator command-plane envelope embeds it (PROPOSAL §6.1.3, §6.9.2).
+use ironclaw_product_contracts::operator_service::{
+    RebornLogLevel, RebornLogQueryResponse, RebornOperatorStatusResponse,
+    RebornServiceLifecycleResponse,
+};
 use ironclaw_threads::{SessionThreadRecord, SummaryArtifact, ThreadMessageRecord};
 use ironclaw_turns::{CancelRunResponse, ResumeTurnResponse, RetryTurnResponse, TurnRunState};
 use secrecy::SecretString;
@@ -25,206 +32,6 @@ const OUTBOUND_DELIVERY_TARGET_ID_MAX_BYTES: usize = 512;
 const OUTBOUND_DELIVERY_CHANNEL_MAX_BYTES: usize = 128;
 const OUTBOUND_DELIVERY_DISPLAY_NAME_MAX_BYTES: usize = 256;
 const OUTBOUND_DELIVERY_DESCRIPTION_MAX_BYTES: usize = 1024;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum RebornOperatorStatusState {
-    Ready,
-    Degraded,
-    Blocked,
-    Unsupported,
-    NotConfigured,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum RebornOperatorStatusSeverity {
-    Info,
-    Warning,
-    Critical,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RebornOperatorStatusCheck {
-    pub id: String,
-    pub status: RebornOperatorStatusState,
-    pub severity: RebornOperatorStatusSeverity,
-    pub summary: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub remediation: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RebornOperatorStatusResponse {
-    pub generated_at: DateTime<Utc>,
-    pub overall: RebornOperatorStatusState,
-    pub checks: Vec<RebornOperatorStatusCheck>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum RebornLogLevel {
-    Trace,
-    Debug,
-    Info,
-    Warn,
-    Error,
-}
-
-#[derive(Debug, Clone, Default, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RebornLogQueryRequest {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub limit: Option<u32>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub cursor: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub level: Option<RebornLogLevel>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub target: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub thread_id: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub run_id: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub turn_id: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub tool_call_id: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub tool_name: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub source: Option<String>,
-    #[serde(default)]
-    pub tail: bool,
-    #[serde(default)]
-    pub follow: bool,
-}
-
-impl RebornLogQueryRequest {
-    pub fn set_limit(mut self, limit: u32) -> Self {
-        self.limit = Some(limit);
-        self
-    }
-
-    pub fn set_cursor(mut self, cursor: impl Into<String>) -> Self {
-        self.cursor = Some(cursor.into());
-        self
-    }
-
-    pub fn set_level(mut self, level: RebornLogLevel) -> Self {
-        self.level = Some(level);
-        self
-    }
-
-    pub fn set_target(mut self, target: impl Into<String>) -> Self {
-        self.target = Some(target.into());
-        self
-    }
-
-    pub fn set_thread_id(mut self, thread_id: impl Into<String>) -> Self {
-        self.thread_id = Some(thread_id.into());
-        self
-    }
-
-    pub fn set_run_id(mut self, run_id: impl Into<String>) -> Self {
-        self.run_id = Some(run_id.into());
-        self
-    }
-
-    pub fn set_turn_id(mut self, turn_id: impl Into<String>) -> Self {
-        self.turn_id = Some(turn_id.into());
-        self
-    }
-
-    pub fn set_tool_call_id(mut self, tool_call_id: impl Into<String>) -> Self {
-        self.tool_call_id = Some(tool_call_id.into());
-        self
-    }
-
-    pub fn set_tool_name(mut self, tool_name: impl Into<String>) -> Self {
-        self.tool_name = Some(tool_name.into());
-        self
-    }
-
-    pub fn set_source(mut self, source: impl Into<String>) -> Self {
-        self.source = Some(source.into());
-        self
-    }
-
-    pub fn set_tail(mut self, tail: bool) -> Self {
-        self.tail = tail;
-        self
-    }
-
-    pub fn set_follow(mut self, follow: bool) -> Self {
-        self.follow = follow;
-        self
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RebornLogEntry {
-    pub id: String,
-    pub timestamp: DateTime<Utc>,
-    pub level: RebornLogLevel,
-    pub target: String,
-    pub message: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub thread_id: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub run_id: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub turn_id: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub tool_call_id: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub tool_name: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub source: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RebornLogQueryResponse {
-    pub source: String,
-    pub entries: Vec<RebornLogEntry>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub next_cursor: Option<String>,
-    pub tail_supported: bool,
-    pub follow_supported: bool,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum RebornServiceLifecycleAction {
-    Install,
-    Start,
-    Stop,
-    Status,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum RebornServiceLifecycleState {
-    Installed,
-    Running,
-    Stopped,
-    Unsupported,
-    Failed,
-    Unknown,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RebornServiceLifecycleRequest {
-    pub action: RebornServiceLifecycleAction,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RebornServiceLifecycleResponse {
-    pub action: RebornServiceLifecycleAction,
-    pub state: RebornServiceLifecycleState,
-    pub message: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub remediation: Option<String>,
-}
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 pub struct RebornChannelConnectAction {
