+++ crates/ironclaw_product_contracts/src/lib.rs
@@ -42,7 +42,9 @@ pub mod error;
 pub mod inbound;
 pub mod interaction_commands;
 pub mod lifecycle_service;
+pub mod llm_config;
 pub mod operator_llm;
+pub mod operator_service;
 pub mod operator_tools;
 pub mod outbound;
 pub mod package_lifecycle;
