+++ crates/ironclaw_reborn_composition/src/llm_admin/mod.rs
@@ -1,3 +1,2 @@
-pub(crate) mod nearai_login_serve;
 pub(crate) mod nearai_mcp;
 pub(crate) mod openai_compat_serve;
